Fix screenshot folder creation bug on windows
If the screenshot path is given with unix style path separators, as it probably should, (e.g. 'test/screenshot/:browser_:viewport_:timestamp.png') the determined path separator from require('path').sep will still be \ on windows systems.

This means that in the following split(pathSep) and most of the following code will not work, because the path contains only / but the separator is \.

Normalizing the path before any further operations fixes this.